CODE RUB: Restructure Models Folder
