+++ OpenAI.NET.Tests.Acceptance/Clients/Completions/CompletionClientTests.Prompts.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Force.DeepCloner;
 using Newtonsoft.Json;
-using OpenAI.NET.Models.Completions;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 using WireMock.RequestBuilders;
 using WireMock.ResponseBuilders;
 using Xunit;
