+++ OpenAI.NET.Tests.Acceptance/Clients/Completions/CompletionClientTests.cs
@@ -5,9 +5,9 @@
 using System;
 using System.Linq;
 using OpenAI.NET.Clients.OpenAIs;
-using OpenAI.NET.Models.Completions;
 using OpenAI.NET.Models.Configurations;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 using Tynamix.ObjectFiller;
 using WireMock.Server;
 
@@ -34,7 +34,7 @@ namespace OpenAI.NET.Tests.Acceptance.Clients.Completions
             };
 
             this.openAIClient = new OpenAIClient(openAiConfiguration);
-        }     
+        }
 
         private static ExternalCompletionRequest ConvertToCompletionRequest(Completion completion)
         {
