+++ OpenAI.NET.Tests.Unit/Services/Foundations/Completions/CompletionServiceTests.Exceptions.Prompt.cs
@@ -6,9 +6,9 @@ using System;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Moq;
-using OpenAI.NET.Models.Completions;
-using OpenAI.NET.Models.Completions.Exceptions;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.Completions.Exceptions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 using RESTFulSense.Exceptions;
 using Xunit;
