+++ OpenAI.NET.Tests.Unit/Services/Foundations/Completions/CompletionServiceTests.Logics.Prompt.cs
@@ -7,8 +7,8 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Force.DeepCloner;
 using Moq;
-using OpenAI.NET.Models.Completions;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 using Xunit;
 
 namespace OpenAI.NET.Tests.Unit.Services.Foundations.Completions
