+++ OpenAI.NET.Tests.Unit/Services/Foundations/Completions/CompletionServiceTests.cs
@@ -9,8 +9,8 @@ using System.Linq.Expressions;
 using KellermanSoftware.CompareNetObjects;
 using Moq;
 using OpenAI.NET.Brokers.OpenAIs;
-using OpenAI.NET.Models.Completions;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 using OpenAI.NET.Services.Foundations.Completions;
 using RESTFulSense.Exceptions;
 using Tynamix.ObjectFiller;
