+++ OpenAI.NET/Brokers/OpenAIs/OpenAIBroker.Completions.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------
 
 using System.Threading.Tasks;
-using OpenAI.NET.Models.ExternalCompletions;
+using OpenAI.NET.Models.Services.Foundations.ExternalCompletions;
 
 namespace OpenAI.NET.Brokers.OpenAIs
 {
