+++ OpenAI.NET/Clients/Completions/ICompletionsClient.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using OpenAI.NET.Models.Completions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
 
 namespace OpenAI.NET.Clients.Completions
 {
