+++ OpenAI.NET/Models/Services/Foundations/Completions/CompletionRequest.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Generic;
 
-namespace OpenAI.NET.Models.Completions
+namespace OpenAI.NET.Models.Services.Foundations.Completions
 {
     public class CompletionRequest
     {
