+++ OpenAI.NET/Models/Services/Foundations/Completions/Exceptions/CompletionDependencyValidationException.cs
@@ -4,7 +4,7 @@
 
 using Xeptions;
 
-namespace OpenAI.NET.Models.Completions.Exceptions
+namespace OpenAI.NET.Models.Services.Foundations.Completions.Exceptions
 {
     internal class CompletionDependencyValidationException : Xeption
     {
