+++ OpenAI.NET/Models/Services/Foundations/Completions/Exceptions/NullCompletionException.cs
@@ -4,7 +4,7 @@
 
 using Xeptions;
 
-namespace OpenAI.NET.Models.Completions.Exceptions
+namespace OpenAI.NET.Models.Services.Foundations.Completions.Exceptions
 {
     public class NullCompletionException : Xeption
     {
