+++ OpenAI.NET/Models/Services/Foundations/ExternalCompletions/ExternalUsage.cs
@@ -2,12 +2,19 @@
 // Copyright (c) Coalition of the Good-Hearted Engineers 
 // ---------------------------------------------------------------
 
-namespace OpenAI.NET.Models.Completions
+using Newtonsoft.Json;
+
+namespace OpenAI.NET.Models.Services.Foundations.ExternalCompletions
 {
-    public class Usage
+    public class ExternalUsage
     {
+        [JsonProperty(propertyName: "prompt_tokens")]
         public int PromptTokens { get; set; }
+
+        [JsonProperty(propertyName: "completion_tokens")]
         public int CompletionTokens { get; set; }
+
+        [JsonProperty(propertyName: "total_tokens")]
         public int TotalTokens { get; set; }
     }
 }
