+++ OpenAI.NET/Services/Foundations/Completions/CompletionService.Exceptions.cs
@@ -4,8 +4,8 @@
 
 using System;
 using System.Threading.Tasks;
-using OpenAI.NET.Models.Completions;
-using OpenAI.NET.Models.Completions.Exceptions;
+using OpenAI.NET.Models.Services.Foundations.Completions;
+using OpenAI.NET.Models.Services.Foundations.Completions.Exceptions;
 using RESTFulSense.Exceptions;
 
 namespace OpenAI.NET.Services.Foundations.Completions
